986 create useLayerHook
