+++ frontend/src/components/MapView/Map/index.tsx
@@ -31,11 +31,10 @@ import {
   StaticRasterLayer,
   WMSLayer,
 } from 'components/MapView/Layers';
+import useLayersHook from 'hook/useLayersHook';
 
 interface MapComponentProps {
   setIsAlertFormOpen: Dispatch<SetStateAction<boolean>>;
-  boundaryLayerId: string;
-  selectedLayers: LayerType[];
   panelHidden: boolean;
 }
 
@@ -44,17 +43,17 @@ type LayerComponentsMap<U extends LayerType> = {
 };
 
 const MapComponent = memo(
-  ({
-    setIsAlertFormOpen,
-    boundaryLayerId,
-    selectedLayers,
-    panelHidden,
-  }: MapComponentProps) => {
+  ({ setIsAlertFormOpen, panelHidden }: MapComponentProps) => {
+    // To remove before merging
+    // eslint-disable-next-line no-console
+    console.log('MapComponent');
     const {
       map: { boundingBox, minZoom, maxZoom, maxBounds },
     } = appConfig;
 
     const dispatch = useDispatch();
+
+    const { selectedLayers, boundaryLayerId } = useLayersHook();
 
     const selectedMap = useSelector(mapSelector);
